requirements.txt
discord.py==1.7.1
python-dotenv==0.17.1
PyGithub==1.54.1